pyproject.toml
[project]
name = "streamlit-test-evan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "db-dtypes>=1.4.4",
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-core>=2.5.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pandas-gbq>=0.31.1",
    "plotly>=6.5.0",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "streamlit>=1.52.1",
    "streamlit-option-menu>=0.4.0",
    "tqdm>=4.67.1",
]
